refactor: :recycle: better segregation of concerns
